Making sure the utility_setup is ran only once
I suspect that this is what's causing the spikes in energy meters when
the integration is reloaded. We do see the hilo_energy sensors creating
duplicates and there might be a link here. We'll need some time to test
this adequatly.

This is to address issue #35 which can't be closed until we're sure.