Changing the way we keep track of state for CostSensor
Apparently, there's a race condition in hass where, when we set the
state for a static entity, by the time the state is updated, it's
reverted back to the original state. This causes the entity state to
bounce back and forth.

When we control what exactly what can actually update the state by
listening to events and only setting it under specific condition, we get
stable results.